requirements.txt
fastapi[standard]~=0.115.5
pydantic-settings~=2.6.1
SQLAlchemy~=2.0.36
psycopg2-binary~=2.9.10
alembic~=1.14.0
bcrypt~=4.2.0
pyjwt~=2.9.0